containers: check for socket name with fqdn
this adds a second check in case where the socket name of both daemon
mon/mds would have been created the socket with the fqdn in the name
instead of the shortname.
Indeed, sometimes it may causes issue according to the hostname
configuration of the node.

There has been a first fix in ceph/ceph-container#895 but it won't be enough
to ensure backward compatibility with OSP deployment prior that patch.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1541303

Signed-off-by: Guillaume Abrioux <gabrioux@redhat.com>